+++ bindgen-tests/tests/headers/dont-wrap-fn-ptr-fields.h
@@ -1,0 +1,5 @@
+// bindgen-flags: --dont-wrap-fn-ptr-fields
+
+typedef struct foo {
+    int (*bar)();
+} foo;